Add files via upload
This miniature compiler "just" mimics original working of compilers. For this to work just hit com.cpp, use the mouse to browse over to ' Compile a Programme ' option type in the name of test file (i have included 4 test files already i.e. a1,a2,a3,a4) and see it in action.
WARNING this is just a basic compiler which mimics somewhat action of c++ compiler.
And yes to use this you should have dosbox and turbo c++, which i know is awkward but in indian schools we used this only :(